+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Added the search icon at the left
   - Moved the spinner of `busy` state to the left.
 - New css helper `.ola-black-bolds`
+- `className` property to  `PanelFooter`
 
 ### Removed
 - The css variable `--inner-gap`